+++ examples/01_HelloWorld.py
@@ -1,4 +1,0 @@
-print("hello world")
-a = 1
-# Double-clicked, this script ends with the pydblclick pause menu.
-# Try <i> : the interactive console knows the variable 'a'.