+++ examples/01a_HelloWorld.py
@@ -1,0 +1,13 @@
+
+
+# No import line on the FIRST line of the script
+# The pydblclick features are only active if pydblclick is registered to run python files :
+#  "py -m pydblclick register" + .py and .pyw extensions manually associated with pydblclick application
+
+
+
+
+print("hello world")
+a = 1
+# Thanks to pydblclick features, when double-clicked, this script ends with the pydblclick pause menu.
+# Try <i> : the interactive console knows the variable 'a' which can be inspected for debugging.