+++ examples/02_Unexpected_exception.py
@@ -1,3 +1,0 @@
-print("hello world")
-print(1 / 0)  # without pydblclick, this traceback would flash away mysteriously
-print("goodbye world")