+++ examples/02a_Unexpected_exception.py
@@ -1,0 +1,5 @@
+
+print("hello world")
+a = 0
+print(1 / a)  # without pydblclick, this traceback would flash away mysteriously
+print("goodbye world")