+++ examples/20b_NoPause_option_but_exception.py
@@ -1,3 +1,4 @@
+import pydblclick  # optional: pip install pydblclick -- or delete this line
 # Don't pause this script in console (unless there is an exception):
 pydblclick_customizations['must_pause_in_console'] = False